+++ lib-typhoon/Icecave/Duct/TypeCheck/Exception/UnexpectedInputException.php
@@ -1,0 +1,11 @@
+<?php
+namespace Icecave\Duct\TypeCheck\Exception;
+
+abstract class UnexpectedInputException extends \InvalidArgumentException
+{
+    public function __construct($message, \Exception $previous = null)
+    {
+        parent::__construct($message, 0, $previous);
+    }
+
+}