+++ lib-typhoon/Icecave/Duct/TypeCheck/Validator/Icecave/Duct/Detail/TokenTypeTypeCheck.php
@@ -1,0 +1,6 @@
+<?php
+namespace Icecave\Duct\TypeCheck\Validator\Icecave\Duct\Detail;
+
+class TokenTypeTypeCheck extends \Icecave\Duct\TypeCheck\AbstractValidator
+{
+}