+++ lib-typhoon/Icecave/Duct/TypeCheck/Validator/Icecave/Duct/Exception/ParserExceptionTypeCheck.php
@@ -1,0 +1,6 @@
+<?php
+namespace Icecave\Duct\TypeCheck\Validator\Icecave\Duct\Exception;
+
+class ParserExceptionTypeCheck extends \Icecave\Duct\TypeCheck\AbstractValidator
+{
+}