+++ lib/Icecave/Duct/Detail/Exception/ParserException.php
@@ -1,6 +1,7 @@
 <?php
-namespace Icecave\Duct\Exception;
+namespace Icecave\Duct\Detail\Exception;
 
+use Icecave\Duct\Exception\SyntaxExceptionInterface;
 use RuntimeException;
 
 class ParserException extends RuntimeException implements SyntaxExceptionInterface