+++ test/suite/Icecave/Duct/Detail/TokenStreamParserTest.php
@@ -1,6 +1,7 @@
 <?php
-namespace Icecave\Duct;
+namespace Icecave\Duct\Detail;
 
+use Phake;
 use PHPUnit_Framework_TestCase;
 use stdClass;
 
@@ -8,7 +9,7 @@ class TokenStreamParserTest extends PHPUnit_Framework_TestCase
 {
     public function setUp()
     {
-        $this->parser = new TokenStreamParser;
+        $this->parser = Phake::partialMock(__NAMESPACE__ . '\TokenStreamParser');
     }
 
     protected function createTokens(array $tokens)
@@ -102,12 +103,23 @@ class TokenStreamParserTest extends PHPUnit_Framework_TestCase
      */
     public function testParse(array $tokens, $expectedValue)
     {
+        $values = array();
+
+        $this->parser->on(
+            'document',
+            function ($value) use (&$values) {
+                $values[] = $value;
+            }
+        );
+
         $tokens = $this->createTokens($tokens);
-        $values = $this->parser->parse($tokens);
-        $this->assertInstanceOf('Icecave\Collections\Vector', $values);
-        $this->assertSame(1, $values->size());
-        $this->assertSame(gettype($expectedValue), gettype($values->back()));
-        $this->assertEquals($expectedValue, $values->back());
+        $this->parser->feed($tokens);
+        $this->parser->finalize();
+
+        $this->assertTrue(is_array($values));
+        $this->assertSame(1, count($values));
+        $this->assertSame(gettype($expectedValue), gettype(end($values)));
+        $this->assertEquals($expectedValue, end($values));
     }
 
     public function parseData()
@@ -137,7 +149,90 @@ class TokenStreamParserTest extends PHPUnit_Framework_TestCase
             array(array('{', 'k1', ':', '{', '}', '}'),                      (object) array('k1' => new stdClass)),
             array(array('{', 'k1', ':', '{', 'k2', ':', '{', '}', '}', '}'), (object) array('k1' => (object) array('k2' => new stdClass))),
             array(array('{', 'k1', ':', '[', 1, ',', 2, ']', '}'),           (object) array('k1' => array(1, 2))),
+        );
+    }
 
+    /**
+     * @dataProvider eventData
+     */
+    public function testParseEvents(array $tokens, $expectedEvents)
+    {
+        $tokens = $this->createTokens($tokens);
+        $this->parser->feed($tokens);
+        $this->parser->finalize();
+
+        $verifiers = array();
+        foreach ($expectedEvents as $eventArguments) {
+            $verifiers[] = call_user_func_array(
+                array(Phake::verify($this->parser), 'emit'),
+                $eventArguments
+            );
+        }
+
+        call_user_func_array(
+            'Phake::inOrder',
+            $verifiers
+        );
+    }
+
+    public function eventData()
+    {
+        return array(
+            array(array(1),                                                  array(array('value', array(1)))),
+            array(array(1.1),                                                array(array('value', array(1.1)))),
+            array(array(true),                                               array(array('value', array(true)))),
+            array(array(false),                                              array(array('value', array(false)))),
+            array(array(null),                                               array(array('value', array(null)))),
+            array(array('foo'),                                              array(array('value', array('foo')))),
+
+            array(
+                array('[', ']'),
+                array(
+                    array('array-open'),
+                    array('array-close'),
+                ),
+            ),
+
+            array(
+                array('[', 1, ',', 2, ',', 3, ']'),
+                array(
+                    array('array-open'),
+                    array('value', array(1)),
+                    array('value', array(2)),
+                    array('value', array(3)),
+                    array('array-close'),
+                ),
+            ),
+
+            array(
+                array('[', '{', '}', ']'),
+                array(
+                    array('array-open'),
+                    array('object-open'),
+                    array('object-close'),
+                    array('array-close'),
+                ),
+            ),
+
+            array(
+                array('{', '}',),
+                array(
+                    array('object-open'),
+                    array('object-close'),
+                ),
+            ),
+
+            array(
+                array('{', 'k1', ':', 1, ',', 'k2', ':', 2, '}',),
+                array(
+                    array('object-open'),
+                    array('object.key', array('k1')),
+                    array('value', array(1)),
+                    array('object.key', array('k2')),
+                    array('value', array(2)),
+                    array('object-close'),
+                ),
+            )
         );
     }
 }