+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Duct Changelog
 
-### 0.1.0
+### 0.2.0 (2013-07-08)
+
+* **[NEW]** Added `EventedParser`, a SAX-JS/Clarinet-like event-based JSON parser
+* **[BC]** Moved `Lexer` and `TokenStreamParser` into `Detail` namespace - these classes a no longer part of the public API
+
+### 0.1.0 (2013-05-02)
 
 * Initial release